--- a/.github/workflows/update-ips.yml
+++ b/.github/workflows/update-ips.yml
@@ -10,9 +10,9 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Set up Go
-        uses: actions/setup-go@v1
+        uses: actions/setup-go@v4
         with:
-          go-version: 1.18beta1
+          go-version: stable
         id: go
 
       - name: Checkout