--- a/go.mod
+++ b/go.mod
@@ -1,10 +1,14 @@
 module github.com/psanford/awsip
 
-go 1.18
-
-require golang.org/x/tools v0.1.8
+go 1.21
 
 require (
+	github.com/gaissmai/cidrtree v0.5.0
+	golang.org/x/tools v0.1.8
+)
+
+require (
+	github.com/gaissmai/extnetip v0.4.0 // indirect
 	golang.org/x/mod v0.5.1 // indirect
 	golang.org/x/sys v0.0.0-20211019181941-9d821ace8654 // indirect
 	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect