Use an efficient data structure for lookups

The goal was always to fix the data structure used here. The tool now
has some use outside of the cli tool so it made sense to do this now.

The package we're using implements a prefix lookup with a treap.

The Benchmark results show a much more consistent and faster lookup
time:

```
goos: linux
goarch: arm64
pkg: github.com/psanford/awsip
                              │     before     │                  after                  │
                              │     sec/op     │     sec/op      vs base                 │
Lookup/2a05:d07f:e0ff::ffff-8   32116.0n ± ∞ ¹     190.5n ± ∞ ¹        ~ (p=1.000 n=1) ²
Lookup/54.74.0.27-8             26983.0n ± ∞ ¹     277.5n ± ∞ ¹        ~ (p=1.000 n=1) ²
Lookup/2a05:d03a:8000::1-8      30954.0n ± ∞ ¹     317.6n ± ∞ ¹        ~ (p=1.000 n=1) ²
Lookup/100.23.255.254-8           5.907n ± ∞ ¹   316.600n ± ∞ ¹        ~ (p=1.000 n=1) ²
Lookup/57.180.0.0-8             26993.0n ± ∞ ¹     300.0n ± ∞ ¹        ~ (p=1.000 n=1) ²
Lookup/127.0.0.12-8             32582.0n ± ∞ ¹     380.4n ± ∞ ¹        ~ (p=1.000 n=1) ²
Lookup/10.48.20.96-8            35668.0n ± ∞ ¹     311.9n ± ∞ ¹        ~ (p=1.000 n=1) ²
Lookup/8.8.8.8-8                32779.0n ± ∞ ¹     325.8n ± ∞ ¹        ~ (p=1.000 n=1) ²
Lookup/2606:4700:4700::1111-8   31530.0n ± ∞ ¹     326.4n ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                           11.99µ           300.7n        -97.49%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```